requirements.txt
aiohttp==3.6.2
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.9.2
fabric==2.5.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Cache==0.13.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
idna==2.9
importlib-metadata==1.6.0
invoke==1.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-jsonapi==0.23.1
multidict==4.7.6
numpy==1.18.4
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.10
telepot==12.7
tzlocal==2.1
urllib3==1.25.9
Werkzeug==0.16.0
xlrd==1.2.0
yarl==1.4.2
zipp==3.1.0
xlsxwriter==1.2.9
flask_cors==3.0.8
reportlab~=3.5.50
openpyxl~=3.0.5
Pillow~=7.2.0
PyJWT==1.7.1
psycopg2-binary==2.8.6
redis==2.10.6
pandas==1.2.0